Create README.md
I've never cided before just downloaded git hub because was looking to aqcuire a better private search engine and stellar vpn. This seems to be totally different then what was advertised when comparing a lust if private search engines.  Well will stick around out if curiosity now and to understand more on coding.